job-interview-ai-agent/requirements.txt
python-dotenv>=1.0.0
openai>=1.12.0
pypdf>=4.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
gradio>=4.19.0
